==> scripts/02_Simulation_n06.Rmd <==
---
output: github_document
---

<!-- 02_Simulation_n06.md is generated from 02_Simulation_n06.Rmd. Please edit that file -->

```{r, include = FALSE}
knitr::opts_chunk$set(
  collapse = TRUE,
  comment = "#>"
)
```

# Introduction

<!-- badges: start -->

<!-- badges: end -->

I want to run a simulation for n=6 taxa. There are 15 possible quadruples with n=6 taxa, and I want to test all possible combinations for k out of 15 quadruples for phylogenetic decisiveness.

For each random set of quadruples, I test both the 4-way-partition-property using the No-Rainbow-Coloring algorithm (**NRC**, truth) and the Fixing Taxon Traceability algorithm (**FTT**, test), both implemented in the R package **FixingTaxonTraceR**. 

# Initialize

I use a file named *SourceFile.R* that contains all relevant R packages and user-/server-specific path to the R library. If using this code, you must make all the necessary changes within the template source file. 

```{r init}
rm(list = ls())
time0<-Sys.time()

source("../SourceFile.R")
source("../helperFunctions/SimulationFunction_exact.R")

```

# Get input data

In *myTab_n6*, there are all possible Four-way-partitions (4WP) given n=6 taxa. 

In *quadruple_data*, there are all possible quadruples given n=6 taxa. 

```{r input}
# Load initial input data
test1 = FTT_createInput(fn="../data/S6_Decisive.txt",sepSym = "_",c = 4)

# Sanity check of FTT package
test_FTT = FTT_algorithmRed(data = test1$data,verbose = T, c=4, n=6)
test_NRC = FTT_findNRC(data = test1)

load("../results/01_partitions/partitions_n06.RData")

```

# Test-Loop with 10 combinations per k

To test less $k$, I use as lower bound the minimal triple covering, $\frac{1}{4}\binom{n}{3}$. As upper bound I use $\binom{n}{4} - (n-4) -1$, as above this $k$ all sets are phylogenetic decisive. 

```{r testloop}
test1$data[,status:=NA]
n=6
LowerBound = choose(n,3)/4
UpperBound = choose(n,4) - n + 3

dumTab = foreach(j=c(LowerBound:UpperBound))%do%{
  # j=5
  message("\nWorking on n=6, k=",j)
  time1 = Sys.time()
  myTest = SimulationFunction_exact(number_taxa = 6, 
                                    number_quads = j,
                                    repeats = 10,
                                    data1 = test1,
                                    data2 = myTab_n06,
                                    verbose = T)
  time2 = Sys.time()
  x0 = as.numeric(round(difftime(time2,time1,units = "mins"),3))
  message("       Total time for n=6, k=",j," & rep = 10: " ,round(difftime(time2,time1,units = "mins"),3)," minutes")
  
  outfn = paste0("../temp/02_SimulationsData_n06/SimulationResults_n6_k",j,".RData")
  save(myTest,file = outfn)
  
  tab = table(myTest$FWPP,myTest$FTT)
  x2 = myTest[FWPP == "PhyloDec" & FTT == "FTT",.N]
  x3 = myTest[FWPP == "NOT PhyloDec",.N]
  x4 = myTest[FWPP == "PhyloDec",.N]
  
  message("       There were ",x2," of ",x4," sets identified by Fischers algorith as decisive (",round(x2/x4,4)*100,"%)")
  
  res = data.table(k = j,
                   time = x0,
                   NR_NotPhyloDec = x3,
                   NR_PhyloDec = x4,
                   NR_FTT = x2)
  res
}
SimulationResults_n06_test = rbindlist(dumTab)

SimulationResults_n06_test[,posRate := NR_FTT/NR_PhyloDec]
SimulationResults_n06_test[NR_PhyloDec==0,posRate := NA]
SimulationResults_n06_test[NR_PhyloDec>0,]

```

# Loop with all combinations

I use for n=6 still the exact test, as the maximal combination number is 6435 < 10000. I test only within a reasonable range

```{r simloop, eval=F}
dumTab = foreach(j=c(LowerBound:UpperBound))%do%{
  # j=5
  message("\nWorking on n=6, k=",j)
  time1 = Sys.time()
  myTest = SimulationFunction_exact(number_taxa = 6, 
                                    number_quads = j,
                                    repeats = 10000,
                                    data1 = test1,
                                    data2 = myTab_n06,
                                    verbose = T)
  time2 = Sys.time()
  x0 = as.numeric(round(difftime(time2,time1,units = "mins"),3))
  message("       Total time for n=6, k=",j," & rep = 100: " ,round(difftime(time2,time1,units = "mins"),3)," minutes")
  
  outfn = paste0("../results/02_SimulationsData_n06/SimulationResults_n6_k",j,"_rep10000.RData")
  save(myTest,file = outfn)

  tab = table(myTest$FWPP,myTest$FTT)
  x2 = myTest[FWPP == "PhyloDec" & FTT == "FTT",.N]
  x3 = myTest[FWPP == "NOT PhyloDec",.N]
  x4 = myTest[FWPP == "PhyloDec",.N]
  
  message("       There were ",x2," of ",x4," sets identified by Fischers algorith as decisive (",round(x2/x4,4)*100,"%)")
  
  res = data.table(k = j,
                   time = x0,
                   NR_NotPhyloDec = x3,
                   NR_PhyloDec = x4,
                   NR_FTT = x2)
  res
}
SimulationResults_n06 = rbindlist(dumTab)

SimulationResults_n06[,posRate := NR_FTT/NR_PhyloDec]
SimulationResults_n06[NR_PhyloDec==0,posRate := NA]
save(SimulationResults_n06, file = "../results/02_SimulationResults_n06.RData")

SimulationResults_n06[NR_PhyloDec>0,]

```


# Session Info

```{r sessionInfo}
sessionInfo()
message("\nTOTAL TIME : " ,round(difftime(Sys.time(),time0,units = "mins"),3)," minutes")

```
