==> scripts/04_MinimalSetSizes.Rmd <==
---
title: "Minimal decisive sets"
author: "Janne Pott"
date: "27/04/2023"
output: github_document
---

<!-- 04_MinimalSetSizes.md is generated from 04_MinimalSetSizes.Rmd. Please edit that file -->

```{r, include = FALSE}
knitr::opts_chunk$set(
  collapse = TRUE,
  comment = "#>"
)
```

# Introduction

<!-- badges: start -->

<!-- badges: end -->

In our publication, we give lower bounds for each simulated $k_n$ for which phylogenetic decisive sets are known. 

Here, I load example sets with the respective $k_n$ and test for both decisiveness using the _findNRC_ algorithm of Parvini et al. and our _fixingTaxonTraceR_ algorithm. 

# Initialize

I use a file names *SourceFile.R* that contains all relevant R packages and user-/server-specific path to the R library. If using this code, you must make all the necessary changes within the template source file. 

```{r init}
rm(list = ls())
time0<-Sys.time()

source("../SourceFile.R")

```

# Load data

The example data is stored within the simulation repository. For $n=6$, $n=7$ and $n=8$, the examples are explicitly also given in the publication (Examples 3.10 and 3.23). The set for $n=9$ was taken from the simulation with $k_9=46$, were exactly one set was decisive. We tested if one additional input sample could be removed and indeed, $\{2, 3, 5, 9\}$ was not necessary for covering all 4-way partitions. For $n=10$, we used the Steiner Quadruple System as baseline and added step-wise cross-quadruples which covered the most of the un-covered partitions. 

```{r input}
# load example sets
data_n06 = FTT_createInput(fn = "../data/S6_Decisive_minimal.txt",
                           sepSym = "_")
data_n07 = FTT_createInput(fn = "../data/S7_Decisive_minimal.txt",
                           sepSym = "_")
data_n08 = FTT_createInput(fn = "../data/S8_Decisive_minimal.txt",
                           sepSym = "_")
data_n09 = FTT_createInput(fn = "../data/S9_Decisive_minimal.txt",
                           sepSym = "_")
data_n10 = FTT_createInput(fn = "../data/S10_Decisive_minimal.txt",
                           sepSym = "_")

# load partitions
load("../results/01_partitions/partitions_n06.RData")
load("../results/01_partitions/partitions_n07.RData")
load("../results/01_partitions/partitions_n08.RData")
load("../results/01_partitions/partitions_n09.RData")
load("../results/01_partitions/partitions_n10.RData")


```

# Test partitions

To test for phylogenetic decisiveness, one can simply test if all possible 4-way partitions are covered. All possible partitions for $n=6, ..., 10$ were created in the simulation, and are used to test the 5 minimal sets. 

```{r  partitions}
time3 = Sys.time()

dummy06 = data_n06$data[status=="input",]
dummy07 = data_n07$data[status=="input",]
dummy08 = data_n08$data[status=="input",]
dummy09 = data_n09$data[status=="input",]
dummy10 = data_n10$data[status=="input",]

myTab_n06[,count :=0]
myTab_n07[,count :=0]
myTab_n08[,count :=0]
myTab_n09[,count :=0]
myTab_n10[,count :=0]

for(i in 1:dim(dummy06)[1]){
  #i=1
  myInput = dummy06[i,ctuple]
  filt = grepl(myInput,myTab_n06$allQuads)
  myTab_n06[filt==T,status := "covered"]
  myTab_n06[filt==T,count := count + 1]
  myTab_n06
}
for(i in 1:dim(dummy07)[1]){
  #i=1
  myInput = dummy07[i,ctuple]
  filt = grepl(myInput,myTab_n07$allQuads)
  myTab_n07[filt==T,status := "covered"]
  myTab_n07[filt==T,count := count + 1]
  myTab_n07
}
for(i in 1:dim(dummy08)[1]){
  #i=1
  myInput = dummy08[i,ctuple]
  filt = grepl(myInput,myTab_n08$allQuads)
  myTab_n08[filt==T,status := "covered"]
  myTab_n08[filt==T,count := count + 1]
  myTab_n08
}
for(i in 1:dim(dummy09)[1]){
  #i=1
  myInput = dummy09[i,ctuple]
  filt = grepl(myInput,myTab_n09$allQuads)
  myTab_n09[filt==T,status := "covered"]
  myTab_n09[filt==T,count := count + 1]
  myTab_n09
}
for(i in 1:dim(dummy10)[1]){
  #i=1
  myInput = dummy10[i,ctuple]
  filt = grepl(myInput,myTab_n10$allQuads)
  myTab_n10[filt==T,status := "covered"]
  myTab_n10[filt==T,count := count + 1]
  myTab_n10
}

myTab_n06[,table(count)]
myTab_n07[,table(count)]
myTab_n08[,table(count)]
myTab_n09[,table(count)]
myTab_n10[,table(count)]

message("\nTOTAL TIME for partitions: " ,
        round(difftime(Sys.time(),time3,units = "mins"),3),
        " minutes")

```

For all 5 sets, the minimal count per partition is $1>0$, e.g. all partitions are covered at least once, which gives us phylogenetic decisiveness. 

# Test FTT

All set sizes are below the lower bound of the FTT algorithm. Hence this section is just a sanity check, as none of them can be resolved with fixing taxa. 

```{r FTT}
time1 = Sys.time()

test1_06_FTT = FTT_algorithmRed(data = data_n06$data,n=6,verbose = T)
test1_07_FTT = FTT_algorithmRed(data = data_n07$data,n=7,verbose = T)
test1_08_FTT = FTT_algorithmRed(data = data_n08$data,n=8,verbose = T)
test1_09_FTT = FTT_algorithmRed(data = data_n09$data,n=9,verbose = T)
test1_10_FTT = FTT_algorithmRed(data = data_n10$data,n=10,verbose = T)

message("\nTOTAL TIME for FTT: " ,
        round(difftime(Sys.time(),time1,units = "mins"),3),
        " minutes")

```

# Test NRC

As all sets are decisive, there should only be rainbow-coloring possible, e.g. regardless how you try to color the $n$ taxa with 4 colors, there is at least one quadruple with 4 colors. This is a direct result of the 4-way partition property: each partition is one possible coloring, and all partitions are covered at least once. 

Please note: the *FTT_findNRC* algorithm is an R adaption of Parvini et al. *findNRC* algorithm implemented in Python. The run time is not optimized. 

```{r NRC}
time2 = Sys.time()

test2_06_NRC = FTT_findNRC(data = data_n06)
test2_07_NRC = FTT_findNRC(data = data_n07)
test2_08_NRC = FTT_findNRC(data = data_n08)
test2_09_NRC = FTT_findNRC(data = data_n09)
test2_10_NRC = FTT_findNRC(data = data_n10)

message("\nTOTAL TIME for NRC: " ,
        round(difftime(Sys.time(),time2,units = "mins"),3),
        " minutes")

```

# Session Info

```{r sessionInfo}
sessionInfo()
message("\nTOTAL TIME : " ,
        round(difftime(Sys.time(),time0,units = "mins"),3),
        " minutes")

```
